requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# SSH/Network
netmiko>=4.3.0
paramiko>=3.4.0

# Structured Output Parsing
ntc-templates>=6.0.0      # Pre-built TextFSM templates for network devices
textfsm>=1.1.3            # TextFSM parser (usually installed with netmiko)
ttp>=0.9.5                # Template Text Parser (simpler alternative)

# Config & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# Templating
jinja2>=3.1.2

# CLI
typer>=0.9.0
rich>=13.7.0

# Development
python-dotenv>=1.0.0
httpx>=0.26.0  # For testing

# Optional: Background tasks
# celery>=5.3.0
# redis>=5.0.0
